cs_policy_interface/requirements.txt
adal>=0.6.0
boto3>=1.13.11
botocore>=1.16.11
google-api-core>=1.21.0
google-api-python-client>=1.12.8
google-auth>=1.23.0
google-auth-httplib2>=0.0.3
google-cloud-bigquery>=1.24.0
jsonschema>=3.2.0
netaddr>=0.7.18
numpy>=1.16.6
oauth2client>=4.1.3
pandas>=0.24.0
pymongo>=3.11.4
pyodbc>=3.0.10
pyotp>=2.3.0
python-dateutil>=2.8.1
pytz>=2013.6
PyYAML>=3.1.0
requests!=2.9.0,>=2.8.1 # Apache-2.0
six>=1.15.0 # MIT
sqlparse>=0.2.3
yaql>=1.1.0
